solucion de errores

+++ database/migrations/2020_12_18_062424_create_mensaje_table.php
@@ -15,8 +15,8 @@ class CreateMensajeTable extends Migration
     {
         Schema::create('mensaje', function (Blueprint $table) {
             $table->bigIncrements('id_mensaje');
-            $table->integer('user_id_envia');
-            $table->integer('user_id_recibe');
+            $table->unsignedBigInteger('user_id_envia');
+            $table->unsignedBigInteger('user_id_recibe');
             $table->dateTime('fecha');
             $table->string('estado', 25);
             $table->string('mensaje', 255);
